fix(convert_source_descriptions): fix linting issues
