+++ convert_source_description/tests/TEST_README.md
@@ -16,6 +16,7 @@ convert_source_description/
 │   ├── __init__.py
 │   ├── TEST_README.md
 │   ├── test_convert_source_description.py
+│   ├── test_stripping_utils.py
 │   └── test_utils_helper.py
 ├── requirements.txt
 └── pytest.ini
@@ -57,6 +58,15 @@ The current fixtures are defined inside `test_utils_helper.py`:
 This file exists as a placeholder for future tests of `convert_source_description.py`.
 It currently contains no active tests.
 
+### Stripping Utils Tests (`test_stripping_utils.py`)
+
+Tests for `stripping_utils.py` currently cover:
+- `strip_by_delimiter()`
+- `strip_tag_and_clean()`
+- `strip_tag()`
+
+The test module validates delimiter handling, tag removal behavior, and paragraph cleanup handling for malformed and nested HTML snippets.
+
 ### Utils Helper Tests (`test_utils_helper.py`)
 
 Tests for `utils_helper.py` currently cover:
@@ -65,7 +75,5 @@ Tests for `utils_helper.py` currently cover:
 - paragraph-content extraction behavior
 - paragraph-index lookup behavior
 - glyph replacement behavior
-- delimiter-based splitting behavior
-- tag stripping and cleanup behavior
 
 The test module uses inline HTML fixtures built with BeautifulSoup to validate parser behavior against realistic source-description fragments.
