+++ convert_source_description/tests/test_utils_helper.py
@@ -495,291 +495,3 @@ class TestReplaceGlyphs:
         actual_result = helper._replace_glyphs(input_string)
         # Assert
         assert actual_result == expected_result
-
-
-class TestStripByDelimiter:
-    """Tests for the _strip_by_delimiter function."""
-
-    @pytest.mark.parametrize(
-        "input_string, delimiter, expected_result",
-        [
-            ("a, b, c", ",", ["a", "b", "c"]),
-            ("a; b; c", ";", ["a", "b", "c"]),
-            ("a: b: c", ":", ["a", "b", "c"]),
-            ("a| b| c", "|", ["a", "b", "c"]),
-        ],
-    )
-    def test_with_single_occurence(
-        self, helper, input_string, delimiter, expected_result
-    ):
-        """Test that the input string with single occurcence is stripped correctly."""
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter(input_string, delimiter)
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    @pytest.mark.parametrize(
-        "input_string, delimiter, expected_result",
-        [
-            ("a ,b ,, c", ",", ["a", "b", "", "c"]),
-            ("a ;b ;; c", ";", ["a", "b", "", "c"]),
-            ("a :b :: c", ":", ["a", "b", "", "c"]),
-            ("a |b || c", "|", ["a", "b", "", "c"]),
-        ],
-    )
-    def test_with_multiple_occurences(
-        self, helper, input_string, delimiter, expected_result
-    ):
-        """Test that the input string with multiple occurences is stripped correctly."""
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter(input_string, delimiter)
-
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    @pytest.mark.parametrize(
-        "input_string, delimiter, expected_result",
-        [
-            ("a, b; c", ";", ["a, b", "c"]),
-            ("a, b; c", ",", ["a", "b; c"]),
-            ("a, b; c", ":", ["a, b; c"]),
-            ("a, b; c", "|", ["a, b; c"]),
-        ],
-    )
-    def test_with_different_delimiters(
-        self, helper, input_string, delimiter, expected_result
-    ):
-        """Test that the input string is stripped correctly with different delimiters."""
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter(input_string, delimiter)
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_leading_delimiter(self, helper):
-        """Test that the input string with leading delimiter is stripped correctly."""
-        # Arrange
-        expected_result = ["", "a", "b", "c"]
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter(",a, b, c", ",")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_trailing_delimiter(self, helper):
-        """Test that the input string with trailing delimiter is stripped correctly."""
-        # Arrange
-        expected_result = ["a", "b", "c", ""]
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter("a, b, c,", ",")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    @pytest.mark.parametrize(
-        "input_string, delimiter, expected_result",
-        [
-            ("a, b, c", ",", ["a", "b", "c"]),
-            (" a ,b, c ", ",", ["a", "b", "c"]),
-            ("   a   ,   b   ,   c   ", ",", ["a", "b", "c"]),
-        ],
-    )
-    def test_with_whitespace(self, helper, input_string, delimiter, expected_result):
-        """Test that the input string with whitespace is stripped correctly."""
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter(input_string, delimiter)
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_delimiter_not_found(self, helper):
-        """Test that it returns the input string when the delimiter is not found."""
-        # Arrange
-        expected_result = ["abcd"]
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter("abcd", ",")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_empty_string_input(self, helper):
-        """Test that it returns an empty list when the input string is empty."""
-        # Arrange
-        expected_result = [""]
-        # Act
-        actual_stripped_result = helper._strip_by_delimiter("", ",")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-
-class TestStripTagAndClean:
-    """Tests for the _strip_tag_and_clean function."""
-
-    def test_valid_tags(self, helper):
-        """Test stripping valid tags and cleaning content."""
-        # Arrange
-        content = "<td><p>Test Content</p></td>"
-        tag = "td"
-        expected_result = "Test Content"
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_nested_tags(self, helper):
-        """Test stripping nested tags and cleaning content."""
-        # Arrange
-        content = "<td><p><strong>Strong Nested Content</strong></p></td>"
-        tag = "td"
-        expected_result = "<strong>Strong Nested Content</strong>"
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_multiple_paragraphs(self, helper):
-        """Test replacing multiple <p> tags with <br />."""
-        # Arrange
-        content = "<td><p>First paragraph</p><p>Second paragraph</p></td>"
-        tag = "td"
-        expected_result = "First paragraph <br /> Second paragraph"
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_multiple_paragraphs_with_nested_tags(self, helper):
-        """Test replacing multiple <p> tags with <br />."""
-        # Arrange
-        content = "<td><p><strong>First paragraph</strong></p><p><em>Second paragraph</em></p></td>"
-        tag = "td"
-        expected_result = (
-            "<strong>First paragraph</strong> <br /> <em>Second paragraph</em>"
-        )
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_malformed_tags(self, helper):
-        """Test content with malformed tags."""
-        # Arrange
-        content = "<td><p><strong>Malformed Content</p></td>"
-        tag = "td"
-        expected_result = "<strong>Malformed Content"
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_no_matching_tags(self, helper):
-        """Test content with no matching tags."""
-        # Arrange
-        content = "<p>No matching tags here</p>"
-        tag = "td"
-        expected_result = ""
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_empty_content(self, helper):
-        """Test empty content."""
-        # Arrange
-        content = ""
-        tag = "td"
-        expected_result = ""
-        # Act
-        result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert result == expected_result
-
-    def test_with_none(self, helper):
-        """Test that it returns an empty string when passed None."""
-        # Arrange
-        content = None
-        tag = "strong"
-        expected_result = ""
-        # Act
-        actual_stripped_result = helper._strip_tag_and_clean(content, tag)
-        # Assert
-        assert actual_stripped_result == expected_result
-
-
-class TestStripTag:
-    """Tests for the _strip_tag function."""
-
-    def test_with_p(self, helper):
-        """Test that it strips a paragraph tag."""
-        # Arrange
-        content = "<p>Hello, World!</p>"
-        expected_result = "Hello, World!"
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "p")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_div(self, helper):
-        """Test that it strips a div tag."""
-        # Arrange
-        content = "<div> This is a div </div>"
-        expected_result = "This is a div"
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "div")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_h1(self, helper):
-        """Test that it strips a heading tag."""
-        # Arrange
-        content = "<h1> Heading 1 </h1>"
-        expected_result = "Heading 1"
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "h1")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_nested_tag(self, helper):
-        """Test that it strips a nested tag."""
-        # Arrange
-        content = "<p><strong>This is a nested tag</strong></p>"
-        expected_result = "<strong>This is a nested tag</strong>"
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "p")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_empty_tag(self, helper):
-        """Test that it returns an empty tag when passed an empty tag."""
-        # Arrange
-        content = "<p></p>"
-        expected_result = ""
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "p")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_plain_text(self, helper):
-        """Test that it returns the plain text when passed plain text."""
-        # Arrange
-        content = "This is a plain text"
-        expected_result = "This is a plain text"
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "p")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_empty_string(self, helper):
-        """Test that it returns an empty string when passed an empty string."""
-        # Arrange
-        content = ""
-        expected_result = ""
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "p")
-        # Assert
-        assert actual_stripped_result == expected_result
-
-    def test_with_none(self, helper):
-        """Test that it returns an empty string when passed None."""
-        # Arrange
-        content = None
-        expected_result = ""
-        # Act
-        actual_stripped_result = helper._strip_tag(content, "p")
-        # Assert
-        assert actual_stripped_result == expected_result
